New test script for trialling tDCS
Added a new version of the test script stimulate_3D_2 which applies iou
current as in the VERTEX website tutorials as well as the tDCS current
(tDCS on it’s own won’t elicit APs). Also fixed the if statement call
in invitroSliceStim and suppressed an uneccessary output in
groupUpdateSchedule.