AC stim fixes and Cat model
Basically copied and modified the sf files to make the cat vis model,
as they are v. similar, and also made fixed to the mirror estimate for
time dependent stim. There is also a new version of tutorial 2 for slow
oscillations instead of Gamma, just with a few minor parameter tweaks.